Wstępne działanie reduxa
