+++ src/App.tsx
@@ -3,19 +3,19 @@ import './App.css';
 import { Game } from './components/game/Game';
 
 import { store, unsubscribe } from './redux/store';
-import { addTodo, changeVisibility } from './redux/actions';
+import { addTodo, changeVisibility, SET_VISIBILITY_FILTER } from './redux/actions';
 
 
 function App() {
-  console.log(store.getState());
+  // console.log(store.getState());
+  
+  // store.subscribe(store.getState());
 
   store.dispatch(addTodo('Hello !!!'));
-  store.dispatch(addTodo('Hola !!!'));
-  store.dispatch(addTodo('Guten morgen !!!'));
-  store.dispatch(addTodo('Siemanko !!!'));
-
-
-  store.dispatch(changeVisibility("myfilter"));
+  // store.dispatch(addTodo('Hola !!!'));
+  // store.dispatch(addTodo('Guten morgen !!!'));
+  // store.dispatch(addTodo('Siemanko !!!'));
+  store.dispatch(changeVisibility(SET_VISIBILITY_FILTER));
 
   unsubscribe();
